+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Add `sdwan_service_routing_ospfv3_ipv4_profile_parcel` resource and data source
 - Add `sdwan_service_routing_ospfv3_ipv6_profile_parcel` resource and data source
 - Add `sdwan_transport_route_policy_profile_parcel` resource and data source
+- Add `sdwan_service_route_policy_profile_parcel` resource and data source
 
 ## 0.3.13